Final Version, some Evaluations

+++ result.h
@@ -9,7 +9,7 @@ struct StepResult
 	unsigned int amount_prey;
 	unsigned int amount_plants;
 	unsigned int current_step;
-	unsigned int operations;
+	unsigned long operations;
 	struct StepResult *next;
 };
 
@@ -17,7 +17,7 @@ struct SimulationResult
 {
 	int simulation_steps;
 	struct StepResult *first_step_result;
-	unsigned int operations;
+	unsigned long operations;
 	struct timeval start_time;
 	struct timeval finish_time;
 	struct timeval run_time;
